Refactor Count down latch to module
