+++ src/lib.rs
@@ -1,4 +1,5 @@
 pub mod cache;
+pub mod count_down_latch;
 pub mod dispatcher;
 pub mod exchanger;
 pub mod execution_limiter;
